:boom:  4K STOP

+++ src/include/lib/printk.h
@@ -47,7 +47,7 @@ struct position
 	int XCharSize;
 	int YCharSize;
 
-	unsigned int *FB_addr;
+	unsigned long *FB_addr;
 	unsigned long FB_length;
 
 	spinlock_T printk_lock;
